+++ src/main/java/bokjak/bokjakserver/util/image/ImageFilePathUtils.java
@@ -1,4 +1,4 @@
-package bokjak.bokjakserver.domain.image.service;
+package bokjak.bokjakserver.util.image;
 
 import bokjak.bokjakserver.domain.image.S3SaveDir;
 import java.net.URLDecoder;
@@ -15,16 +15,16 @@ public class ImageFilePathUtils {
     private static final String S3_OBJECT_NAME_PATTERN = "{0}_{1}.{2}";
 
 
-    static String buildRootPath(String bucket, S3SaveDir saveDir) {
+    public static String buildRootPath(String bucket, S3SaveDir saveDir) {
         return bucket + saveDir.path;
     }
 
-    static String buildImageFilePath(MultipartFile multipartFile, String owner) {    // 실제 파일 이름으로부터 key 생성
+    public static String buildImageFilePath(MultipartFile multipartFile, String owner) {    // 실제 파일 이름으로부터 key 생성
         String fileExtension = getFileExtension(Objects.requireNonNull(multipartFile.getOriginalFilename()));
         return MessageFormat.format(S3_OBJECT_NAME_PATTERN, owner, UUID.randomUUID(), fileExtension);
     }
 
-    static String buildImageFilePath(String url) {   // URL로부터 key 생성
+    public static String buildImageFilePath(String url) {   // URL로부터 key 생성
         String[] parsedUrl = url.split(URL_SEPARATOR);
         String fileName = parsedUrl[parsedUrl.length - 1];
         return URLDecoder.decode(fileName, StandardCharsets.UTF_8);
